% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{hint}
\alias{hint}
\title{Give the player a hint for the current question.}
\usage{
hint()
}
\description{
This function will give you a short hint regarding the solution to the
current question.
}
\note{
Viewing a hint costs R Power.
}
\examples{
hint()

}
