% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeGame.R
\name{initializeGame}
\alias{initializeGame}
\title{Initialize the Castle of R game environment}
\usage{
initializeGame(playerLevel)
}
\description{
This function reads in the various text files containing the data and
configuration of the various game components such as rooms, doors, objects
and riddles.
}
