% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{lockDoor}
\alias{lockDoor}
\title{Lock door to another room}
\usage{
lockDoor(n)
}
\arguments{
\item{n}{Door number, as it appears next to the door description, when the
player enters the room, or with the \code{wtf} function}
}
\description{
This function will lock the door to another room, if it is open.
}
\note{
A locked door means another question if you ever wish to go back from
that door to the room from which you came from. However, it also means more
time before Lady R gets into the room.
}
\examples{
lockDoor(1)

}
