% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{openDoor}
\alias{openDoor}
\title{Open door to another room}
\usage{
openDoor(n)
}
\arguments{
\item{n}{Door number}
}
\description{
This function will transfer the player to another room if she answers a
question correctly
}
\examples{
openDoor(1)
}
