% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react.R
\name{react}
\alias{react}
\title{React to player's actions during the Castle of R game}
\usage{
react(game, ...)
}
\description{
This function is called in the background after each of the player's actions
and responds accordingly. It is the "engine" of the game, modifying the game
environment according to the game's logic.
}
