% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Room}
\alias{Room}
\title{R6 classes for the Castle of R game}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Room
}
\description{
R6 classes for the Castle of R game
}
\keyword{datasets}
