% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{seeMap}
\alias{seeMap}
\title{Look at a map}
\usage{
seeMap(n)
}
\arguments{
\item{n}{Floor number (1 for 1st floor, etc.)}
}
\description{
This function lets you look at a map, if you have one. The map is of an
entire floor in the Castle.
}
\examples{
seeMap(1)

}
