% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{seePassword}
\alias{seePassword}
\title{Look at the password}
\usage{
seePassword()
}
\description{
This function will let you look at the current situation of the password.
}
\examples{
seePassword()

}
