% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{solution}
\alias{solution}
\title{Give the player the solution for the current question.}
\usage{
solution()
}
\description{
This function will give you the solution to the current question.
}
\note{
Viewing a solution costs R Power.
}
\examples{
solution()

}
