% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startGame.R
\name{startGame}
\alias{startGame}
\title{Start the Castle of R game}
\usage{
startGame(...)
}
\description{
This function will start the Castle of R game. If you played the game before,
ended the game in the middle and chose to save your progress - this
function will identify this and will give you the option to pick up where
you left. It will also suggest to clean your workspace and plots entirely,
for a better game experience. Be careful before accepting this suggestion.
}
\examples{
startGame()

}
