% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{takeObject}
\alias{takeObject}
\title{Take an object in a room}
\usage{
takeObject(n)
}
\arguments{
\item{n}{Object number, as it appears next to the object description, when the
player enters the room, or with the \code{wtf} function)}
}
\description{
This function lets you take a specific object while in a room.
}
\note{
Some objects might give you valuable clues on how to escape the Castle.
}
\examples{
takeObject(1)

}
