% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{timeLeft}
\alias{timeLeft}
\title{Give the player the solution for the current question.}
\usage{
timeLeft()
}
\description{
This function will give you the solution to the current question.
}
\note{
Viewing a solution costs R Power.
}
\examples{
solution()

}
