% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{whatDoIHave}
\alias{whatDoIHave}
\title{Summarise a player's resources}
\usage{
whatDoIHave()
}
\description{
This function will remind you the current content of your satchel, how much R
Power you have and other resources.
}
\examples{
whatDoIHave()

}
