% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{whatWasTheQuestion}
\alias{whatWasTheQuestion}
\title{Remind a player of the current question}
\usage{
whatWasTheQuestion()
}
\description{
This function will remind you the current question in case you forgot.
}
\examples{
whatWasTheQuestion()

}
