% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{wtf}
\alias{wtf}
\title{Debrief a player of his overall whereabouts}
\usage{
wtf()
}
\description{
This function repeats the greeting message of a room, detailing the doors
and objects (if any) around you. It also summarizes how much R Power you
have and the objects you have taken in your satchel.
}
\examples{
wtf()

}
