% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyFunctions.R
\name{timeLeft}
\alias{timeLeft}
\title{View the time left in a room until Lady R is coming}
\usage{
timeLeft()
}
\description{
This function will print the time until Lady R is coming in a given room.
}
\note{
See note in the \code{lockDoor} function documentation about getting
more time.
}
\examples{
timeLeft()

}
